Added Variant Manager
